energy-anomaly/requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
tensorflow>=2.16
matplotlib>=3.8
pyyaml>=6.0
requests>=2.32
tqdm>=4.66
